Revise API skeleton
