+++ backend/database/database_setup.py
@@ -1,0 +1,5 @@
+import database
+
+db = database.db
+
+db.create_all()
